fix: stage-order defect — dev deps no longer ship in the published analyzer image
- addon/Dockerfile: runtime is now the LAST stage (hassio builder 2026.02.1
  builds the final stage; dev stage previously leaked pytest/ruff/httpx +
  the test suite, ~28 MB amd64 / ~26 MB aarch64)
- addon/Dockerfile.dev (NEW): dev stage split out for Tier 2 iteration only
- addon-build.yml: probe-runtime-image job rebuilds the runtime Dockerfile
  and asserts pytest/ruff/httpx/tests are ABSENT (acceptance evidence)
- test_api.py: 422 reason payload asserted (undecodable + empty upload)
- addon/config.yaml: version 1.0.0 -> 1.0.1 (integration manifest stays 2.1.0)
- README: real image sizes (~526 MB amd64 / ~549 MB aarch64) + Deezer
  coverage limitation note
- test_tempo.py: RUF046 lint fix (pre-existing)

+++ analyzer/tests/test_tempo.py
@@ -68,7 +68,7 @@ class TestTempo:
         assert abs(bpm - 90.0) <= 2.0
 
     def test_silence_returns_none(self):
-        silence = np.zeros(int(44100 * 10), dtype=np.float32)
+        silence = np.zeros(44100 * 10, dtype=np.float32)
         assert tempo.estimate_bpm(silence) is None
 
     def test_too_short_returns_none(self):
